// webui/src/services/types/enums.ts
// 枚举类型定义
export enum Mode {
  Free = 'Free',
  Collaborate = 'Collaborate',
  Stage = 'Stage', 
  Event = 'Event',
  Story = 'Story'
}

export enum Event {
  Compete = 'Compete',
  Team = 'Team',
  Tour = 'Tour',
  Challenge = 'Challenge',
  Trial = 'Trial',
  Mission = 'Mission'
}

export enum Choose {
  Loop = 'Loop',
  Random = 'Random',
  ListUp = 'ListUp',
  ListDown = 'ListDown',
  No = 'No'
}

export enum Level {
  Special = 'Special',
  Expert = 'Expert',
  Hard = 'Hard',
  Normal = 'Normal',
  Easy = 'Easy'
}

export enum Performance {
  AllPerfect = 'AllPerfect',
  FullCombo = 'FullCombo',
  Custom = 'Custom',
  DropLastCustom = 'DropLastCustom'
}

export enum FetchMode {
  FetchOne = 'FetchOne',
  FetchLack = 'FetchLack',
  FetchSongHeader = 'FetchSongHeader',
  SpecialChar = 'SpecialChar'
}

export enum WorkflowMode {
  WalkThrough = 'WalkThrough',
  WalkThroughSheet = 'WalkThroughSheet',
  Capture = 'Capture',
  Record = 'Record',
  TraceNote = 'TraceNote',
  TraceFristNote = 'TraceFristNote'
}